Let the Weaponmastery spell buff innate attacks.
Intended for Draconians who pick the metamorphosis power. Currently it
buffs innate attacks and all wielded weapons, but if that is too strong
(on e.g. Centaurs or Beastmen with the right mutations) it could be
changed to work like the rings of power and only buff innate attacks
when the player has no weapons.